♻️ refactor : 거리 요금 계산 로직 타입 캐스팅

+++ src/main/java/nextstep/subway/domain/BasicDistanceFareFormula.java
@@ -31,6 +31,6 @@ public class BasicDistanceFareFormula implements DistanceFareFormula {
         if (distance <= 0) {
             return 0;
         }
-        return (int) ((Math.ceil((distance - 1) / basicAdditionalDistance) + 1) * ADDITIONAL_FARE);
+        return (int) ((Math.ceil((double) (distance - 1) / basicAdditionalDistance) + 1) * ADDITIONAL_FARE);
     }
 }
